move things around, create react component
