+++ package.json
@@ -14,17 +14,19 @@
     "babel-plugin-polyfill-regenerator": "0.0.2",
     "codemirror": "^5.54.0",
     "immer": "^7.0.7",
-    "markdown-toc": "^1.2.0",
     "react": "16.11.0",
     "react-codemirror2": "^7.2.0",
     "react-dom": "16.11.0",
+    "react-markdown": "^4.3.1",
     "redux": "^4.0.5",
     "styled-components": "4.4.1"
   },
   "scripts": {
-    "start": "react-scripts start"
+    "start": "react-scripts start",
+    "generate-spec": "node scripts/generate-spec.js"
   },
   "devDependencies": {
+    "markdown-toc": "^1.2.0",
     "prettier": "~2.0.5",
     "react-scripts": "3.4.1"
   },
