+++ scripts/spec-input.md
